requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cramjam==2.11.0
cycler==0.12.1
et_xmlfile==2.0.0
fastparquet==2024.11.0
fonttools==4.57.0
fsspec==2025.9.0
idna==3.10
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.7
lightgbm==4.6.0
matplotlib==3.9.4
nltk==3.9.1
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
polars==1.32.3
pyarrow==21.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.8.29
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
zipp==3.21.0
